### 2025-12-17 3e47734
Revise affiliation details
