### 2025-12-17 2d93eaf
Update paper.bib with new references
