+++ paper.bib
@@ -160,9 +160,60 @@
   publisher={IEEE}
 }
 
-@misc{benchmark_data,
-  title={Decentralized Internet SDK Performance Benchmarks},
-  author={Lonero Team},
-  year={2024},
-  howpublished={\url{https://github.com/Lonero-Team/Decentralized-Internet/wiki/Benchmarks}}
+@article{adoption_matching,
+  title={An Idea by Lonero That Can Change Millions of Lives},
+  author={Kamal, Andrew Magdy},
+  journal={HackerNoon},
+  year={2020},
+  howpublished={\url{https://hackernoon.com/an-idea-by-lonero-that-can-change-millions-of-lives-ttr3t7x}}
 }
+
+@article{decentralized_caching,
+  title={3.6 Million Websites Went Offline Due to a Fire: Decentralization Prevents That},
+  author={Kamal, Andrew Magdy},
+  journal={HackerNoon},
+  year={2021},
+  howpublished={\url{https://hackernoon.com/36-million-websites-went-offline-due-to-a-fire-decentralization-prevents-that-g42b33um}}
+}
+
+@article{dapp_proposal,
+  title={Proposal to Pi Network: Use Our SDK for Your DAPP Development Platform},
+  author={Kamal, Andrew Magdy},
+  journal={HackerNoon},
+  year={2021},
+  howpublished={\url{https://hackernoon.com/proposal-to-pi-network-use-our-sdk-for-your-dapp-development-platform-h2j337r}}
+}
+
+@article{genetic_optimization,
+  title={Genetic Algorithm Optimization Using Distributed Computing},
+  author={Kamal, Andrew Magdy},
+  journal={engrXiv Preprints},
+  year={2020},
+  doi={10.31224/osf.io/preprints/engrxiv/1508},
+  howpublished={\url{https://engrxiv.org/index.php/engrxiv/preprint/view/1508}}
+}
+
+@article{biostatistical_analysis,
+  title={A Distributed Architecture Proposal for Regressional Generative Adversarial Networks for Biostatistical Analysis Modeled after the Decentralized-Internet SDK and BOIN},
+  author={Andrew Magdy Kamal Nassief},
+  journal={Semantic Scholar},
+  year={2019},
+  howpublished={\url{https://pdfs.semanticscholar.org/9115/9c383e10dbc1d40f82df7c495c0d20185e06.pdf}}
+}
+
+@article{postquantum_encryption,
+  title={Distributed Web Systems Leading to Hardware Oriented Cryptography and Post-Quantum Cryptologic Methodologies},
+  author={Kamal, Andrew Magdy and others},
+  journal={IACR Cryptology ePrint Archive},
+  year={2019},
+  number={1467},
+  howpublished={\url{https://eprint.iacr.org/2019/1467.pdf}}
+}
+
+@misc{starkcom_usage,
+  title={Starkcom Global - Decentralized 5G Mesh Networking},
+  author={Starkcom Global},
+  year={2024},
+  howpublished={\url{https://starkcom.io}}
+}
+
