### 2025-12-17 #9639 @danielskatz
👋 @Mentors4EDU

I notice that there are a number of bots that have contributed to the repo - Can you discuss the use of AI, particularly in the context of JOSS's guidance: https://joss.readthedocs.io/en/latest/submitting.html#interim-position-on-ai-enabled-software

Also, given that your paper says "The foundation of the SDK consists of battle-tested open-source libraries", are the source for these packages in the repo?  Can you explain the structure of the repo briefly, highlighting what is new vs what is incorporated from other existing packages?
