### 2025-12-18 #9639 @Mentors4EDU
Hi @danielskatz,

I really appreciate your guidance.

The bots I’ve designated as GitHub contributors are actually GitHub actions specifically designed to manage vulnerabilities and oversee dependency management. This is crucial for an SDK/library of this scale, especially since everything is organized as a monorepo. However, I don’t use them for code generation. When it comes to generative code, less than 5% of the codebase was created using any form of AI—mostly for interfacing with ports, handling regex expressions, and managing dependencies for Theia. No AI-generated code has been used in the first 444 releases on NPM, nor in any of the SDK’s core functionalities.

For licensing management, I rely on FOSSA to monitor all licensing aspects and track dependencies. In terms of battle-testing, aside from the SDK’s growing popularity and wide usage (with over 1800 pulls/downloads from GitHub in the last two weeks), OpenHub reports that the SDK has no significant vulnerabilities. Impressively, there haven’t been any major CVE vulnerabilities reported since 2019, which is quite an accomplishment for software of this size.

Since the repository is a monorepo, I’ve added two main enhancements to make it easier to use beyond what’s described in the documentation: One is a control panel:

Additionally, I provide a development IDE built with Theia. My goal is to enable inter-peer communication through a locally established network. Essentially, there’s a main node or server, and each connected computer operates as a peer, gradually forming a grid. This setup is primarily used for high-performance computing and model training, though I’m working toward simplifying the integration process and establishing a new communication standard.

Ultimately, I want implementation to be as easy as writing a few lines of code, with streamlined APIs and interfaces. My focus is on maintaining a balance i.e. keeping things simple for general users while still accommodating HPC computing and clustering needs. I’m also aiming for inclusion in JOSS to make the code more citable for researchers, since relying only on Zenodo doesn’t provide ideal indexing for research.
