### 2026-01-02 e82c988
chore(deps): update actions/checkout action to v6
