### 2026-01-11 0e17400
Enhance paper with new sections and details
Added sections on software design, AI usage disclosure, research impact, and future development.
