### 2026-01-12 29e16bb
Fix "Mongobleed" vulnerability CVE-2025-14847
