### 2026-01-12 7ce7c66
Fix "Mongobleed" vulnerability CVE-2025-14847
