### 2026-01-15 #9639 @jGaboardi
Regarding [required sections](https://joss.readthedocs.io/en/latest/paper.html#:~:text=Your%20paper%20must%20include%20the%20following%20required%20sections), please:
* Include a *State of the field* section
* Promote *AI Usage Disclosure* to a top-level section
