### 2026-01-15 #9639 @jGaboardi
It seems like my [`check repository`](https://github.com/openjournals/joss-reviews/issues/9639#issuecomment-3752499914) command may be stalled out or simply taking longer usual. However, it does appear that the paper [grew substantially since it's initial submission](https://github.com/openjournals/joss-reviews/issues/9639#issuecomment-3662498150). Once EditorialBot reports back I'll comment more.
