### 2026-01-15 7e01f69
Update paper.bib with additional publication details
Added missing DOI and updated publication details for several entries.
