+++ paper.bib
@@ -23,6 +23,7 @@
   booktitle={International Workshop on Peer-to-Peer Systems},
   pages={53--65},
   year={2002},
+  doi={10.1007/3-540-45748-8_5},
   organization={Springer}
 }
 
@@ -39,6 +40,7 @@
   journal={Fifth IEEE/ACM International Workshop on Grid Computing},
   pages={4--10},
   year={2004},
+  doi={10.1109/grid.2004.14},
   organization={IEEE}
 }
 
@@ -146,6 +148,7 @@
   booktitle={IFIP/ACM International Conference on Distributed Systems Platforms and Open Distributed Processing},
   pages={329--350},
   year={2001},
+  doi={10.1007/3-540-45518-3_18},
   organization={Springer}
 }
 
@@ -157,6 +160,7 @@
   number={2},
   pages={72--93},
   year={2005},
+  doi={10.1109/comst.2005.1610546},
   publisher={IEEE}
 }
 
@@ -185,12 +189,12 @@
 }
 
 @article{genetic_optimization,
-  title={Genetic Algorithm Optimization Using Distributed Computing},
+  title={Distributed Computing for Genetic Approximation Algorithms for the Kissing Problem (n-dimensions)},
   author={Kamal, Andrew Magdy},
   journal={engrXiv Preprints},
-  year={2020},
-  doi={10.31224/osf.io/preprints/engrxiv/1508},
-  howpublished={\url{https://engrxiv.org/index.php/engrxiv/preprint/view/1508}}
+  year={2021},
+  doi={10.31224/osf.io/zva4h},
+  howpublished={\url{https://engrxiv.org/preprint/view/1508}}
 }
 
 @article{biostatistical_analysis,
@@ -198,6 +202,7 @@
   author={Andrew Magdy Kamal Nassief},
   journal={Semantic Scholar},
   year={2019},
+  doi={10.22541/au.157755932.23851366},
   howpublished={\url{https://pdfs.semanticscholar.org/9115/9c383e10dbc1d40f82df7c495c0d20185e06.pdf}}
 }
 
@@ -216,4 +221,3 @@
   year={2024},
   howpublished={\url{https://starkcom.io}}
 }
-
