### 2026-01-15 e10782e
Expand paper
Added sections on the state of the field and AI usage disclosure.
