+++ paper.md
@@ -39,6 +39,14 @@ The design separates core networking, cryptographic, and coordination primitives
 
 The project favors a lightweight SDK and library model rather than a heavy, opinionated platform, which can mean fewer “batteries-included” features compared to some larger frameworks, but grants tighter control over performance, resource usage, and deployment footprint. When evaluating whether to build or contribute, existing tools like IPFS, libp2p, and other distributed computing frameworks were considered, but many focus narrowly on content addressing, generic P2P networking, or specific blockchain ecosystems rather than providing an end-to-end SDK tailored for decentralized web and distributed compute pipelines as envisioned here. Creating new software made it possible to experiment with a distinct architecture aimed at offloading computation, clustering, and HPC-style workloads, while still remaining interoperable where useful. This matters because it offers developers a focused, customizable foundation for building next-generation decentralized applications and compute grids that do not need to conform to the assumptions of existing ecosystems.
 
+# State of the field
+Currently, over the past few decades, there has been an increase in cryptocurrencies since Nakamoto's Bitcoin, which is based on distributed Proof of Work. Within that timeframe, the emergence of BOINC and high-performance computing has paved the way for next-generation computing and highlighted the necessity for fault-tolerant infrastructure across a diverse range of fields. Presently, with the advent of LLMs (Large Language Models), autonomy in warfare, and the influence of social media, many individuals express concerns regarding the centralization of the internet. Furthermore, issues of censorship and privacy rights have escalated in importance. For instance, certain regimes impose censorship on specific religious content, seminaries, and missionaries, while others suppress political dissent. Amidst these pressing concerns, the rise of 'cyberpunks' aiming to establish decentralized infrastructure has not gone unnoticed. This SDK aims to provide distributed computing for training highly data-intensive libraries while also being sufficiently modular to accommodate encrypted communications, privacy applications, and distributed consensus.
+
+# AI Usage Disclosure
+This project uses automation and AI in a limited, carefully controlled way, and this disclosure explains that usage. Designated “bot” contributors in the repository are continuous-integration automations focused on security and maintenance tasks such as vulnerability scanning and dependency management in a large monorepo; they are not used for code generation. Generative AI has been applied to less than 5% of the overall codebase, and only for narrow, well-bounded tasks such as interfacing with specific ports, writing regex utilities, and configuring dependencies for the Theia-based development environment. No AI-generated code appears in the first 444 NPM releases or in any core SDK functionality.
+
+Licensing and dependency compliance are managed with specialized tooling (e.g., FOSSA) rather than generative AI, and security quality is further supported by external monitoring and usage-based validation. Generative AI has not been used to implement fundamental architecture, high-performance computing features, or clustering logic. Instead, it has served as a minor aid around peripheral implementation details, with all AI-assisted contributions reviewed and integrated by a human maintainer. This approach is intended to preserve human accountability for design decisions while transparently acknowledging the narrow, supporting role that AI has played in the project’s development.
+
 # Architecture and Design
 
 The Decentralized Internet SDK follows a modular architecture consisting of several interconnected components:
@@ -71,11 +79,6 @@ The SDK integrates OpenPeer AI Cloud-Agents [@cloudagents] to create a machine l
 - Predictive resource allocation
 - Automated anomaly detection
 - Route analysis and optimization
-
-### AI Usage Disclosure
[…]
