### 2026-01-16 47ec96b
Added DOI for Bitcoin Whitepaper
10.2139/ssrn.3440802
