+++ paper.bib
@@ -31,6 +31,7 @@
   title={Bitcoin: A peer-to-peer electronic cash system},
   author={Nakamoto, Satoshi},
   year={2008},
+  doi={10.2139/ssrn.3440802},
   howpublished={\url{https://bitcoin.org/bitcoin.pdf}}
 }
 
