### 2026-01-21 #9639 @GuilloteauQ
> 👋 [@julienmalard](https://github.com/julienmalard),[@pradeeban](https://github.com/pradeeban), & [@GuilloteauQ](https://github.com/GuilloteauQ) - would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @jGaboardi, I am willing to review this submission (this would be my first for JOSS 🎉 ).
