### 2026-01-21 #9639 @jGaboardi
@GuilloteauQ Great! Thanks for the prompt response. We are excited for your first JOSS review!
