### 2026-01-21 #9639 @julienmalard
I could review.
