### 2026-01-21 #9875 @editorialbot
[REVIEW]: Decentralized Internet SDK: A Comprehensive Framework for Building Peer-to-Peer and Distributed Computing Applications
**Submitting author:** @Mentors4EDU (<a href="http://orcid.org/0000-0001-5389-1386">Andrew Kamal</a>)
**Repository:** https://github.com/Lonero-Team/Decentralized-Internet
**Branch with paper.md** (empty if default branch):
**Version:** v5.4.3
