### 2026-01-21 7041756
chore(deps): update dependency packaging to v26
