+++ clusterpost/bigchaindb/docs/root/requirements.txt
@@ -3,7 +3,7 @@ recommonmark==0.7.1
 sphinx-rtd-theme==3.1.0
 sphinxcontrib-napoleon==0.7
 sphinxcontrib-httpdomain==1.8.1
-packaging==25.0
+packaging==26.0
 pygments==2.19.2 # not directly required, pinned by Snyk to avoid a vulnerability
 setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
 requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
