+++ clusterpost/bigchaindb/docs/server/requirements.txt
@@ -5,7 +5,7 @@ sphinxcontrib-napoleon==0.7
 sphinxcontrib-httpdomain==1.8.1
 pyyaml==6.0.3
 aafigure==0.6
-packaging==25.0
+packaging==26.0
 pygments==2.19.2 # not directly required, pinned by Snyk to avoid a vulnerability
 certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
 setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
