+++ clusterpost/bigchaindb/docker-compose.yml
@@ -8,7 +8,7 @@ services:
   # Build: docker-compose build -d bigchaindb
   # Run: docker-compose run -d bdb
   mongodb:
-    image: mongo:8.2.3
+    image: mongo:8.2.4
     ports:
       - "27017:27017"
     command: mongod
