### 2026-02-01 #9875 @GuilloteauQ
Hello, I am a bit confused here...

The git history is very difficult to navigate...
Most of the recent activity in this repository is done by bots to check/update dependencies. The meaningful contributions seem to be 6 years old imported in a batch.
The other contributions of the author seem to be minor updates to files likes README.md, packages.json, Cargo.toml.
(`git shortlog --author="Andrew Kamal" | awk 'NR>1 && !/snyk/ && !/dependabot/ && !/renovate/ && !/depfu/ { print $0 }'`)

In the paper, all the examples of usage seems to point to things done by the author, or to generic websites where i could not validate the claim.
The paper states that there is a Spack package for this repository, but the package does not work given that it is using a makefile build sequence, but the repository never had a makefile (https://github.com/spack/spack-packages/blob/develop/repos/spack_repo/builtin/packages/decentralized_internet/package.py), and is also 6 years old.

I am stopping here for today, but would love to know what you think @julienmalard !
