### 2026-02-02 #9875 @julienmalard
Hello @GuilloteauQ ,

I agree that most of the meaningful code in the package seems to have been written several years ago. In and of itself, though, I wouldn't see this as a major problem ; I don't see an issue with an article on a mature project being published some time after its development. In the rapidly evolving field of P2P and decentralised technologies, however, this does make me wonder if the package still follows best practices and is up-to-date with recent innovations in the field.

Here are my initial thoughts:
* What I find more concerning is that the package seems to be a "catch-all" for many types of decentralised technologies. While the statement of need is well-written, it's not clear to me exactly why these should all be inside a single repository instead of "pick-and-choose" modules as, for instance, IPFS and libp2p are configured. It is also not clear how this project relates to or complements (or simply uses and builds upon?) such modern systems as IPFS and libp2p.
* The code, as structured, is difficult to navigate.
* The documentation I found linked to the project was for a cryptocurrency called "Lonero," and not for the Decentralized Internet SDK.
* Several comments in the paper are hard to justify in my opinion (e.g., the link between "autonomous warfare" and the problem of a centralised internet - both are real problems, but how will fixing the latter address the former?).
* It is not clear to me how "AI automation" works. "AI" can be any of a large number of approaches; specifically which was used?
* Use of the package for "large-scale distributed computing networks" is claimed, but no examples are given.
* No clear links to tests demonstrating the claimed performance indices are given.
* I was unable to find automated tests for the core functionalities of the package.
* The project seems to be an agglomeration of many different projects, and the links between these, as well as the underlying common thread justifying their presentation as a group, are not at all clear. At the very least, a diagram could help better understand the relationships between these packages...otherwise, perhaps the author could consider splitting the article into discreet articles, each aimed at a unique part of the ecosystem presented.

Overall, I find that the technological questions addressed are very much worthwhile, but that the code and article require very significant improvement and reorganisation.
