### 2026-02-02 #9875 @Mentors4EDU
Hi @GuilloteauQ and @julienmalard,

Thank you both for taking the time to look at my paper and my monorepo. Both of you raise some valid points which are not too difficult to answer.

I will be providing a comprehensive response to some of your questions and concerns soon.
