### 2026-02-04 #9875 @jGaboardi
@GuilloteauQ & @julienmalard – I can see you both have started your reviews with great questions and concerns for @Mentors4EDU.

@Mentors4EDU – Nice timely acknowledgement for your engagement with their comments. Try to have your complete responses to the points they raise in ≈2 weeks time.
