### 2026-02-22 #9875 @Mentors4EDU
Hi @julienmalard and @GuilloteauQ.

Thank you both for your thorough and thoughtful feedback. I greatly appreciate the time and consideration you have taken to review the *Decentralized Internet SDK* and its associated repository.

Regarding the age and activity history of the repository, this project has been under consistent, long-term maintenance. The automated tools you observed, such as those handling dependency and security updates, are an intentional part of the project’s maintenance strategy, ensuring reliability, security, and compatibility with progressing ecosystems.

In practice, these automated workflows enable efficient management of thousands of pull requests over several years, which would be infeasible to maintain manually. This approach reflects established best practices in software engineering, particularly for mature monorepositories.

It is important to note that this repository operates as a **monorepo**, incorporating various interrelated modules within the broader *Decentralized Internet SDK* framework. While this may result in a complex version history, it supports an integrated development environment where advancements in one component strengthen others. This organizational structure is common among large-scale open-source frameworks and should not detract from the project’s eligibility for inclusion in JOSS.

The association with the **Lonero Foundation** that you mentioned reflects the project’s earlier development management and infrastructure, which explains why certain documentation URLs are connected under that domain. The *Lonero* reference is therefore more of a "historical artifact", not a separate or unrelated effort.

In terms of ongoing work, the SDK has continued to progress through recent updates, including the **decentg** package, the **Theia-based development environment**, and its dedicated interface. Although the on-premise and privacy-preserving nature of the software limits analytics on user deployment, community engagement metrics (such as downloads, forks, contributions, and repository activity) indicate sustained interest and potential real-world adoption within the open-source ecosystem.

Though the design is modular with various integrated libraries and components, libraries and monorepos shouldn't be excluded from JOSS publication. The *Decentralized Internet SDK* directly aligns with the mission of JOSS to promote impactful, well-engineered, and accessible open-source research software. Its comprehensive design, ongoing maintenance, and relevance to decentralized computing make it a strong candidate for publication.

I sincerely appreciate your constructive feedback and look forward to your response and the next steps in this review process. Would you like me to provide additional technical context or updated documentation references to assist with your evaluation?
