### 2026-02-22 #9875 @jGaboardi
@GuilloteauQ and @julienmalard – Please respond to @Mentors4EDU's comments in due course.

> ... should not detract from the project’s eligibility for inclusion in JOSS...
> ... directly aligns with the mission of JOSS...
> ... make it a strong candidate for publication.

@Mentors4EDU – While I do appreciate the positive & inclusive sentiment, keep in mind that these decisions are that of the editorial board. Reviewers should come to these conclusions (or not) independently.
