### 2026-03-05 #9875 @jGaboardi
@GuilloteauQ and @julienmalard – Please share your thoughts on the [comments](https://github.com/openjournals/joss-reviews/issues/9875#issuecomment-3939849935) from @Mentors4EDU's as time permits.
