### 2026-03-10 #9875 @julienmalard
Thank you @Mentors4EDU for your initial response; your comments provide important additional context on the projet. For me to recommend publication, however, the initial comments I made above will need to be addressed in the code and in the paper. Some are minor comments, while others will require more effort; one of my main concerns is that the project and code include many seemingly disparate parts that are not clearly conceptually linked to each other (e.g., DAPP, post-quantum encryption, biostatistical computing, 5G mesh computing, blockchain...). These should be clearly linked and streamlined, or else the paper should be reoriented to cover one part of the ecosystem. As it is, it is very hard to navigate both the code as well as the documentation; installation instructions are not clear, and no verifiable examples are given to back up performance and other claims made in the paper. In brief, the code, paper and especially online documentation require significant reworking to effectively highlight the innovativeness and core functionality of the project.
