### 2026-03-19 #9875 @jGaboardi
@Mentors4EDU

So from @julienmalard comments it does seem like some deeply substantive changes will have to happen to the paper and the code base to proceed with review. We will see what @GuilloteauQ thinks and then determine what happens next from there.
