+++ clusterpost/bigchaindb/setup.py
@@ -85,7 +85,7 @@ install_requires = [
     'flask-cors~=6.0.0',
     'flask-restful~=0.3.0',
     'requests>=2.20.0',
-    'gunicorn~=25.1.0',
+    'gunicorn~=25.3.0',
     'jsonschema~=4.26.0',
     'pyyaml>=4.2b1',
     'aiohttp~=3.0',
