+++ clusterpost/bigchaindb/docs/root/requirements.txt
@@ -4,7 +4,7 @@ sphinx-rtd-theme==3.1.0
 sphinxcontrib-napoleon==0.7
 sphinxcontrib-httpdomain==2.0.0
 packaging==26.0
-pygments==2.19.2 # not directly required, pinned by Snyk to avoid a vulnerability
+pygments==2.20.0 # not directly required, pinned by Snyk to avoid a vulnerability
 setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
 requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
 zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
