### 2026-04-02 #9875 @GuilloteauQ
@jGaboardi Yes, I am still willing to. I am currently trying to catch an important deadline this Friday (which took all my time recently...).
I will be able to continue focusing on this JOSS submission on Monday.
