+++ clusterpost/bigchaindb/docs/server/requirements.txt
@@ -6,7 +6,7 @@ sphinxcontrib-httpdomain==2.0.0
 pyyaml==6.0.3
 aafigure==0.6
 packaging==26.0
-pygments==2.19.2 # not directly required, pinned by Snyk to avoid a vulnerability
+pygments==2.20.0 # not directly required, pinned by Snyk to avoid a vulnerability
 certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
 setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
 requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
