### 2026-04-10 #9875 @GuilloteauQ
Here are some of my comments. I do agree with @julienmalard about the state of the repository and the documentation which makes it difficult to evaluate the submission.

# Some points from the checklist

## Development history and open-source practice

- most of the commits are updates of dependencies from bots
```
[Decentralized-Internet]$ git shortlog --summary --numbered --email
  7807  Andrew Kamal <kamalandrew55@gmail.com>
  2367  dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
  2179  Renovate Bot <bot@renovateapp.com>
  1957  renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
   463  dependabot-preview[bot] <27856297+dependabot-preview[bot]@users.noreply.github.com>
   304  snyk-bot <snyk-bot@snyk.io>
   220  depfu[bot] <23717796+depfu[bot]@users.noreply.github.com>
     3  Dilum Aluthge <dilum@aluthge.com>
     2  Loganaden Velvindron <logan@cyberstorm.mu>
     1  David "DijitaL" Bynane <dbynane@gmail.com>
     1  Ikko Eltociear Ashimine <eltociear@gmail.com>
     1  ImgBotApp <ImgBotHelp@gmail.com>
     1  fossabot <badges@fossa.io>
 ```
- I was unable to verify the claim of 175,000 downloads of the SDK in the paper

## Installation

The repository being a mono repo, I seem to understand that the installation process depends on each folder of the repo (i.e., each sub project)
None of the installation step documented in the different readme files worked for me.
(npm cannot install, the spack package mentioned in the paper is not working)

Spack package: https://github.com/spack/spack-packages/blob/develop/repos/spack_repo/builtin/packages/decentralized_internet/package.py

I tried several of the installation proposed here: https://lonero.readthedocs.io/en/latest/Decentralized%20Internet%20Docs/Main%20Installation%20Methods.html

the only one that was working was the docker one (with an image from 6 years ago).

## Documentation

There is a documentation page on the "lonero" page:
https://lonero.readthedocs.io/en/latest/Decentralized%20Internet%20Docs/Critical%20Components.html

I found the documentation confusing and I was not able to follow and understand what was being shown.

## Functionality

I was unable to confirm the functionality of the SDK

## Paper

Summary:
- The paper does contain an accessible description of the high-level functionalities and goals of the software.
- We could argue that the first and last sentences of the summary are very close in meaning/information and does slightly repeat itself.

Statement of Need
- The paper does contain a statement of need section.
- This section however fails to reference other state-of-the-art or state-of-the-practice tools addressing similar, even partial, problems.
- Note that the names of some tools are briefly mentioned in the "Software Design" section.

State of the field:
- The paper does contain a "State of the field" section.
- However, this section does not provide any comparison to any other commonly-used tools/packages.
- Note that the names of some tools are briefly mentioned in the "Software Design" section.
- Instead, the section discusses the way this SDK could help "cyberpucks" establish a decentralized infrastructure to avoid censorship and privacy concerns.

Software Design:
- The paper does contain a "Software Design" section.
- In this section, the author gives the motivation for a modular SDK rather than a larger framework.
- The author discusses the decision to build the SDK from scratch compared to contributing to other existing tools. The decision to build the SDK was taken because the other tools were each focusing on a too narrow application.

Research Impact Statement:
- The paper does contain a "Research Impact Statement".
- In this section, the author claims that the SDK was used and had "demonstrated real-world impact", however no source was given.
- The remaining of the section, describes the high-level advantages of the SDK, and its accessibility through the Spack package manager, which can be used by researchers to install/use the SDK on HPC clusters.
- There is an "example usage" section

Example Usage:
- Each example of the list points to different webpages, but they do not support the fact that the SDK has been used in those contexts
- some of those links point to blog posts and article written by the author
- I was unable to verify the claims of the "Community and Impact" section

AI usage disclosure:
- The paper does contain an AI usage disclosure section.
- The author mentions the use of CI tools for checking security and maintenance tasks.
- The author claims that less than 5% of the code base has been written by AI, and no AI was used in the architecture design of in the critical components of the SDK.
- There are no mention of the use of AI for the generation of the documentation, nor for the authoring of this paper.

Quality of writing:
- While the paper sounds well written, the content sections are a bit scattered and the arguments could be improved.

# Notes for the author

I am still a bit confused with the architecture of the repository which is difficult to navigate.
The documentation is unclear, and the installation commands not working.
I would appreciate it if the author could provide a working example and improve the documentation in order for us reviewers to attest of the functionality of the SDK.
