### 2026-04-12 #9875 @jGaboardi
@Mentors4EDU

Please take some time to review and digest the comments from @julienmalard[^1] and @GuilloteauQ[^2]. From my read through them, they are pretty much on the same page in terms of *major, substantive* changes that would need to happen in terms of the code base, documentation (particularly installation & example usage), and paper. Once you have given this thorough consideration, please respond with:

1. Whether you are willing to perform the rather significant tasks associated with the reviewers' comments; and
2. If are willing to put forth that effort, please present a clear and concise plan for addressing their concerns.

Once we have your answers for *1.* and *2.*, we will go from there.

[^1]: https://github.com/openjournals/joss-reviews/issues/9875#issuecomment-4028546873
[^2]: https://github.com/openjournals/joss-reviews/issues/9875#issuecomment-4223625547
