### 2026-04-26 #9875 @jGaboardi
@Mentors4EDU  re: https://github.com/openjournals/joss-reviews/issues/9875#issuecomment-4232531708

Please respond to both *1.* and *2.* above. If you have not responded within the next 2 weeks, I will likely need to start the recommended rejection process with the Track EiC[^1].

cc @danielskatz

[^1]: As a note to myself – I am also sending this message via email.
