### 2026-05-24 #9875 @Mentors4EDU
Hi, I appreciate the response. I'm still working on addressing their concerns per my previous response. The core issues at play is:
- Modularity: Working on making sure everything plays together in an easier to understand, well documented way.
- Age of Documentation and Clarity: Some installation is old and the lonero.readthedocs.io site needs a complete rewrite or better yet a website dedicated solely to this SDK.
- Use of Automation: As noted before, DevSecOps efforts and the use of tools like dependabot are for managing security vulnerabilities and keeping packages maintained. Perhaps, more can be highlighted on the paper.md and it can link to a separate security.md file to cypher this in greater detail.
- Real World Usage: This is privacy-preserving and most decentralized grids are hosted on-prem, but I will try to come up with better way to explain usage and adoption beyond just the blog posts and screenshots of the stats.
- Usability: Tools to implement this such as a UI, Gateway controller software, and a Theia implementation all exist. This can be documented better however.
- Paper Clarity: Few parts in the paper here and there might seem like a stretch, so I will try to explain things more concisely. These use cases are all possibilities.

Overall, I have been working on maintaining some of the DevSecOps stuff and am making room for a complete overall for some of the docs. This can take roughly 4 to 8 weeks, but I hope to make decent enough progress in the next couple of weeks so this can transition further in the review pipeline.
