### 2026-08-20 #9875 @jGaboardi
> I need this extended out until September 1st

@Mentors4EDU – Do you mean having things complete and ready for re-review by September 1st? If so, then yes this was our [initial agreement - an August 31 deadline](https://github.com/openjournals/joss-reviews/issues/9875#issuecomment-4566239326). If this is not what you mean, then I will enforce that deadline and mark as out of scope.
