### 2026-09-01 #9875 @Mentors4EDU
I think it is best to mark this out of scope and then I can resubmit when it is much more ready or break it down if I want to in the future to potentially smaller items, if I see fit. I don't want to waste JOSS's time and really tried hard, but just don't have the bandwidth to completely restructure this to where it should be in the too intermediate future.
