### 2026-09-01 #9875 @jGaboardi
@Mentors4EDU – Thank you for the honesty.

@julienmalard & @GuilloteauQ – Thank you for your time during this review process.

@danielskatz – At this point I'll be adding the `query scope` tag at the submitter's request to start the "out of scope" process.
